requirements.txt
activity==0.1
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
Appium-Python-Client==1.2.0
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.23.0
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
config==0.5.0.post0
DateTime==4.3
discord.py==1.7.2
form==0.0.1
geojson==2.5.0
idna==2.10
idna-ssl==1.1.0
multidict==5.1.0
pafy==0.5.5
pip==10.0.1
pyowm==2.10.0
pytz==2021.1
requests==2.25.1
selenium==3.141.0
setuptools==39.0.1
soupsieve==2.2.1
tabulate==0.8.9
typing-extensions==3.10.0.0
urllib3==1.26.5
uuid==1.30
yarl==1.6.3
zope.interface==5.4.0